test: extract provider sections to tests/providers/ (batch C, #1440)
