+++ test-all.mjs
@@ -4996,1150 +4996,6 @@ try {
   fail(`custom instructions test crashed: ${e.message}`);
 }
 
-// ── 36. Provider — arbeitnow ────────────────────────────────────
-console.log('\n36. Provider — arbeitnow');
-
-try {
-  const arbeitnowModule = await import(pathToFileURL(join(ROOT, 'providers/arbeitnow.mjs')).href);
-  const arbeitnow = arbeitnowModule.default;
-  const { normalizeArbeitnowJob } = arbeitnowModule;
-
-  if (arbeitnow.id === 'arbeitnow') pass('arbeitnow.id is "arbeitnow"');
-  else fail(`arbeitnow.id is ${JSON.stringify(arbeitnow.id)}`);
-
-  // normalizeArbeitnowJob — field mapping.
-  const full = normalizeArbeitnowJob(
-    { title: '  Staff AI Engineer  ', url: '  https://www.arbeitnow.com/jobs/x1  ', company_name: '  Acme Co  ', location: '  Berlin  ', remote: false, created_at: 1782693032 },
-    'Fallback',
-  );
-  if (full && full.title === 'Staff AI Engineer' && full.url === 'https://www.arbeitnow.com/jobs/x1'
-      && full.company === 'Acme Co' && full.location === 'Berlin' && full.postedAt === 1782693032000) {
-    pass('normalizeArbeitnowJob maps + trims title/url/company/location and converts created_at seconds → ms');
-  } else {
-    fail(`normalizeArbeitnowJob full row = ${JSON.stringify(full)}`);
-  }
-
-  // remote:true appends "Remote" to the location.
-  const remoteJob = normalizeArbeitnowJob({ title: 'R', url: 'https://www.arbeitnow.com/jobs/r', location: 'Munich', remote: true }, 'X');
-  if (remoteJob?.location === 'Munich, Remote') pass('normalizeArbeitnowJob appends "Remote" when remote is true');
-  else fail(`normalizeArbeitnowJob remote location = ${JSON.stringify(remoteJob?.location)}`);
-
-  // remote-only (no base location) → "Remote".
-  const remoteOnly = normalizeArbeitnowJob({ title: 'R', url: 'https://www.arbeitnow.com/jobs/r2', remote: true }, 'X');
-  if (remoteOnly?.location === 'Remote') pass('normalizeArbeitnowJob yields "Remote" when remote is true and location is absent');
-  else fail(`normalizeArbeitnowJob remote-only location = ${JSON.stringify(remoteOnly?.location)}`);
-
-  // company fallbacks: entry name, then "Arbeitnow".
-  const coFromEntry = normalizeArbeitnowJob({ title: 'T', url: 'https://www.arbeitnow.com/jobs/c1', company_name: '' }, 'Entry Name');
-  const coDefault = normalizeArbeitnowJob({ title: 'T', url: 'https://www.arbeitnow.com/jobs/c2' });
-  if (coFromEntry?.company === 'Entry Name' && coDefault?.company === 'Arbeitnow') {
-    pass('normalizeArbeitnowJob falls back company → entry name → "Arbeitnow"');
-  } else {
-    fail(`normalizeArbeitnowJob company fallbacks = ${JSON.stringify({ a: coFromEntry?.company, b: coDefault?.company })}`);
-  }
-
-  // drops: empty title, missing url, non-https url, malformed url, non-object.
-  const drops = [
-    normalizeArbeitnowJob({ title: '', url: 'https://www.arbeitnow.com/jobs/d1' }),
-    normalizeArbeitnowJob({ title: 'No URL' }),
-    normalizeArbeitnowJob({ title: 'Insecure', url: 'http://www.arbeitnow.com/jobs/d3' }),
-    normalizeArbeitnowJob({ title: 'Relative', url: '/jobs/d4' }),
-    normalizeArbeitnowJob({ title: 'Off host', url: 'https://evil.example/jobs/d5' }), // host-lock: external https dropped
-    normalizeArbeitnowJob(null),
-  ];
-  if (drops.every(r => r === null)) pass('normalizeArbeitnowJob drops empty-title / no-url / non-https / relative / off-host / non-object');
-  else fail(`normalizeArbeitnowJob drops = ${JSON.stringify(drops)}`);
-
-  // missing created_at → no postedAt key.
-  const noDate = normalizeArbeitnowJob({ title: 'T', url: 'https://www.arbeitnow.com/jobs/nd' });
-  if (noDate && !('postedAt' in noDate)) pass('normalizeArbeitnowJob omits postedAt when created_at is absent');
-  else fail(`normalizeArbeitnowJob postedAt presence = ${JSON.stringify(noDate)}`);
-
-  // fetch(): pagination by self-built ?page=N, stop on a short page.
-  const mk = (i) => ({ title: `Role ${i}`, url: `https://www.arbeitnow.com/jobs/x${i}`, company_name: `Co ${i}`, location: 'Berlin', remote: false, created_at: 1782693032 + i });
-  const page1 = Array.from({ length: 100 }, (_, i) => mk(i));            // full page → continue
-  const page2 = [mk(100), mk(101), { title: '', url: 'https://www.arbeitnow.com/jobs/bad' }]; // short page (3 < 100) → stop; 1 drop
-  const requested = [];
-  const pagedFetch = async (url, opts) => {
-    requested.push({ url, redirect: opts?.redirect });
-    const u = new URL(url);
-    const page = Number(u.searchParams.get('page'));
-    // links.next deliberately points at a featured ?search= URL to prove we DON'T follow it.
-    if (page === 1) return { data: page1, links: { next: 'https://www.arbeitnow.com/api/job-board-api?search=foo&page=2' }, meta: {} };
-    if (page === 2) return { data: page2, links: { next: null }, meta: {} };
-    return { data: [], links: {}, meta: {} };
-  };
-
-  const paged = await arbeitnow.fetch({ name: 'Arbeitnow' }, { fetchJson: pagedFetch });
-
-  if (requested.length === 2
-      && requested[0].url === 'https://www.arbeitnow.com/api/job-board-api?page=1'
-      && requested[1].url === 'https://www.arbeitnow.com/api/job-board-api?page=2') {
-    pass('arbeitnow.fetch() builds ?page=N URLs itself (ignores links.next) and stops after the short page');
-  } else {
-    fail(`arbeitnow.fetch() requested = ${JSON.stringify(requested.map(r => r.url))}`);
-  }
-
-  if (requested.every(r => r.redirect === 'error')) pass('arbeitnow.fetch() passes redirect:"error" on every page (SSRF guard)');
-  else fail(`arbeitnow.fetch() redirect opts = ${JSON.stringify(requested.map(r => r.redirect))}`);
-
-  if (paged.length === 102) pass('arbeitnow.fetch() aggregates valid jobs across pages (100 + 2, dropping the empty-title row)');
-  else fail(`arbeitnow.fetch() returned ${paged.length} jobs (expected 102)`);
-
-  // max_pages cap: only the first page is requested even though it is full.
-  const capRequested = [];
-  await arbeitnow.fetch(
-    { name: 'Arbeitnow', max_pages: 1 },
-    { fetchJson: async (url, opts) => { capRequested.push(url); return { data: Array.from({ length: 100 }, (_, i) => mk(i)) }; } },
-  );
-  if (capRequested.length === 1 && capRequested[0] === 'https://www.arbeitnow.com/api/job-board-api?page=1') {
-    pass('arbeitnow.fetch() honors max_pages (stops at the cap even on a full page)');
-  } else {
-    fail(`arbeitnow.fetch() max_pages:1 requested ${JSON.stringify(capRequested)}`);
-  }
-
-  // unexpected API response → throws.
-  let badThrew = false;
-  try {
-    await arbeitnow.fetch({ name: 'X' }, { fetchJson: async () => ({ wrong: true }) });
-  } catch (e) {
-    badThrew = /unexpected API response/.test(e.message);
-  }
-  if (badThrew) pass('arbeitnow.fetch() throws on unexpected API response shape');
-  else fail('arbeitnow.fetch() should throw when the data array is absent');
-
-} catch (e) {
-  fail(`arbeitnow provider tests crashed: ${e.message}`);
-}
-
-// ── 37. Provider — rippling ─────────────────────────────────────
-console.log('\n37. Provider — rippling');
-
-try {
-  const ripplingModule = await import(pathToFileURL(join(ROOT, 'providers/rippling.mjs')).href);
-  const rippling = ripplingModule.default;
-  const { parseRipplingResponse } = ripplingModule;
-
-  if (rippling.id === 'rippling') pass('rippling.id is "rippling"');
-  else fail(`rippling.id is ${JSON.stringify(rippling.id)}`);
-
-  // detect(): ats.rippling.com/<slug>/jobs → board API URL.
-  const hit = rippling.detect({ name: 'Acme', careers_url: 'https://ats.rippling.com/acme-corp/jobs' });
-  if (hit && hit.url === 'https://api.rippling.com/platform/api/ats/v1/board/acme-corp/jobs') {
-    pass('rippling.detect() resolves ats.rippling.com/<slug>/jobs → board API URL');
-  } else {
-    fail(`rippling.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  // detect() also works when careers_url is just /<slug> (no /jobs suffix).
-  const hitNoJobs = rippling.detect({ name: 'X', careers_url: 'https://ats.rippling.com/acme-corp' });
-  if (hitNoJobs && hitNoJobs.url === 'https://api.rippling.com/platform/api/ats/v1/board/acme-corp/jobs') {
-    pass('rippling.detect() derives the slug from the first path segment (no /jobs needed)');
-  } else {
-    fail(`rippling.detect() no-/jobs returned ${JSON.stringify(hitNoJobs)}`);
-  }
-
-  if (rippling.detect({ name: 'X', careers_url: 'https://example.com/acme/jobs' }) === null) {
-    pass('rippling.detect() returns null for non-rippling hosts');
-  } else {
-    fail('rippling.detect() should return null for non-rippling hosts');
-  }
-
-  // careers_url with non-string value → detect() returns null without crashing.
-  if (rippling.detect({ name: 'X', careers_url: null }) === null && rippling.detect({ name: 'X', careers_url: 7 }) === null) {
-    pass('rippling.detect() returns null for non-string careers_url (null and 7)');
-  } else {
-    fail('rippling.detect() should treat non-string careers_url as missing');
-  }
-
-  // SSRF/format: non-https, empty path (no slug), and host-spoof in the path.
-  if (rippling.detect({ name: 'X', careers_url: 'http://ats.rippling.com/acme/jobs' }) === null
-      && rippling.detect({ name: 'X', careers_url: 'https://ats.rippling.com/' }) === null
-      && rippling.detect({ name: 'X', careers_url: 'https://evil.example/ats.rippling.com/acme/jobs' }) === null) {
-    pass('rippling.detect() rejects non-https, empty-path, and path-spoofed URLs');
-  } else {
-    fail('rippling.detect() must reject non-https / empty-path / path-spoofed URLs');
-  }
-
-  // Slug safety: a first segment that is not a clean token (space, dot, hyphen-edged) is rejected.
-  if (rippling.detect({ name: 'X', careers_url: 'https://ats.rippling.com/a%20b/jobs' }) === null
-      && rippling.detect({ name: 'X', careers_url: 'https://ats.rippling.com/acme.corp/jobs' }) === null
-      && rippling.detect({ name: 'X', careers_url: 'https://ats.rippling.com/-acme/jobs' }) === null) {
-    pass('rippling.detect() rejects unsafe slugs (space, dot, leading hyphen)');
-  } else {
-    fail('rippling.detect() must reject unsafe slugs');
-  }
-
-  // Internal hyphens are valid.
-  if (rippling.detect({ name: 'X', careers_url: 'https://ats.rippling.com/just-appraised-jobs/jobs' })?.url
-      === 'https://api.rippling.com/platform/api/ats/v1/board/just-appraised-jobs/jobs') {
-    pass('rippling.detect() accepts slugs with internal hyphens');
-  } else {
-    fail('rippling.detect() should accept internal hyphens in the slug');
-  }
-
-  // parseRipplingResponse — deterministic sample (top-level array).
-  const sample = [
-    { uuid: '1', name: 'Account Executive', url: 'https://ats.rippling.com/acme/jobs/uuid-1', department: { label: 'Sales' }, workLocation: { label: 'Remote (United States)', id: 'x' } },
-    { uuid: '2', name: '  ML Engineer  ', url: '  https://ats.rippling.com/acme/jobs/uuid-2  ', workLocation: { label: 'Canada' } },
-    { uuid: '3', name: 'String Loc Role', url: 'https://ats.rippling.com/acme/jobs/uuid-3', workLocation: 'New York' }, // workLocation as bare string
-    { uuid: '4', name: 'No Loc Role', url: 'https://ats.rippling.com/acme/jobs/uuid-4', workLocation: null },           // null → ''
-    { uuid: '5', name: '', url: 'https://ats.rippling.com/acme/jobs/uuid-5' },                                          // drop: empty name
-    { uuid: '6', name: 'No URL Role' },                                                                                 // drop: no url
-    { uuid: '7', name: 'Insecure', url: 'http://ats.rippling.com/acme/jobs/uuid-7' },                                   // drop: non-https
-  ];
-  const jobs = parseRipplingResponse(sample, 'Acme');
-
-  if (jobs.length === 4) pass('parseRipplingResponse keeps 4 valid postings (drops empty-name / no-url / non-https)');
-  else fail(`parseRipplingResponse returned ${jobs.length} postings (expected 4)`);
-
-  if (jobs[0] && Object.keys(jobs[0]).sort().join(',') === 'company,location,title,url') {
-    pass('parseRipplingResponse returns the normalized { title, url, company, location } shape');
-  } else {
-    fail(`parseRipplingResponse row 0 keys = ${JSON.stringify(jobs[0] && Object.keys(jobs[0]))}`);
-  }
-
-  if (jobs[0]?.title === 'Account Executive'
-      && jobs[0]?.url === 'https://ats.rippling.com/acme/jobs/uuid-1'
-      && jobs[0]?.company === 'Acme'
-      && jobs[0]?.location === 'Remote (United States)') {
-    pass('parseRipplingResponse maps name→title, url, company from entry name, workLocation.label→location');
-  } else {
-    fail(`parseRipplingResponse row 0 = ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[1]?.title === 'ML Engineer' && jobs[1]?.url === 'https://ats.rippling.com/acme/jobs/uuid-2') {
-    pass('parseRipplingResponse trims whitespace from name and url');
-  } else {
-    fail(`parseRipplingResponse row 1 title/url = ${JSON.stringify({ title: jobs[1]?.title, url: jobs[1]?.url })}`);
-  }
-
-  if (jobs[2]?.location === 'New York' && jobs[3]?.location === '') {
-    pass('parseRipplingResponse accepts a bare-string workLocation and yields "" when workLocation is null');
-  } else {
-    fail(`parseRipplingResponse loc fallbacks = ${JSON.stringify({ str: jobs[2]?.location, none: jobs[3]?.location })}`);
-  }
-
-  if (parseRipplingResponse({}, 'X').length === 0 && parseRipplingResponse(null, 'X').length === 0) {
-    pass('parseRipplingResponse: non-array input → empty result (no crash)');
-  } else {
-    fail('parseRipplingResponse should yield empty result for non-array input');
-  }
-
-  // Regression: the per-item url is host-locked to ats.rippling.com — an external
-  // https URL is dropped, a valid ats.rippling.com posting URL is kept.
-  const hostLocked = parseRipplingResponse(
-    [
-      { name: 'External Host', url: 'https://evil.example/acme/jobs/uuid-x' },
-      { name: 'Valid Host', url: 'https://ats.rippling.com/acme/jobs/uuid-9' },
-    ],
-    'Acme',
-  );
-  if (hostLocked.length === 1 && hostLocked[0]?.title === 'Valid Host'
-      && hostLocked[0]?.url === 'https://ats.rippling.com/acme/jobs/uuid-9') {
-    pass('parseRipplingResponse host-locks the posting url to ats.rippling.com (drops external https URLs)');
-  } else {
-    fail(`parseRipplingResponse host-lock = ${JSON.stringify(hostLocked)}`);
-  }
-
-  // fetch(): requests the derived API URL and passes the SSRF guard.
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await rippling.fetch(
-    { name: 'Acme', careers_url: 'https://ats.rippling.com/acme-corp/jobs' },
-    { fetchJson: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://api.rippling.com/platform/api/ats/v1/board/acme-corp/jobs') {
-    pass('rippling.fetch() requests the derived board API URL');
-  } else {
-    fail(`rippling.fetch() requested ${JSON.stringify(capturedUrl)}`);
-  }
-
-  if (capturedOpts && capturedOpts.redirect === 'error') {
-    pass('rippling.fetch() passes redirect:"error" to fetchJson (SSRF guard)');
-  } else {
-    fail(`rippling.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-  }
-
-  if (fetched.length === 4 && fetched[0]?.company === 'Acme') {
-    pass('rippling.fetch() returns normalized jobs with company from entry name');
-  } else {
-    fail(`rippling.fetch() returned ${fetched.length} jobs, row 0 = ${JSON.stringify(fetched[0])}`);
-  }
-
-  // fetch(): a non-rippling careers_url cannot derive an endpoint → throws.
-  let badEntryThrew = false;
-  try {
-    await rippling.fetch(
-      { name: 'X', careers_url: 'https://example.com/careers' },
-      { fetchJson: async () => [] },
-    );
-  } catch (e) {
-    badEntryThrew = /cannot derive API URL/.test(e.message);
-  }
-  if (badEntryThrew) pass('rippling.fetch() throws when the careers_url is not an ats.rippling.com host');
-  else fail('rippling.fetch() should throw for a non-rippling careers_url');
-
-} catch (e) {
-  fail(`rippling provider tests crashed: ${e.message}`);
-}
-
-// ── 38. Provider — thehub ───────────────────────────────────────
-console.log('\n38. Provider — thehub');
-
-try {
-  const thehubModule = await import(pathToFileURL(join(ROOT, 'providers/thehub.mjs')).href);
-  const thehub = thehubModule.default;
-  const { normalizeHubJob } = thehubModule;
-
-  if (thehub.id === 'thehub') pass('thehub.id is "thehub"');
-  else fail(`thehub.id is ${JSON.stringify(thehub.id)}`);
-
-  // normalizeHubJob — full mapping.
-  const full = normalizeHubJob(
-    { title: '  Staff Engineer  ', absoluteJobUrl: 'https://thehub.io/jobs/abc123', company: { name: '  Light  ' }, location: { address: '  London, UK  ' }, publishedAt: '2026-06-02T06:59:54.025Z' },
-    'Fallback',
-  );
-  if (full && full.title === 'Staff Engineer' && full.url === 'https://thehub.io/jobs/abc123'
-      && full.company === 'Light' && full.location === 'London, UK'
-      && full.postedAt === Date.parse('2026-06-02T06:59:54.025Z')) {
-    pass('normalizeHubJob maps title/absoluteJobUrl/company.name/location.address + publishedAt → postedAt');
-  } else {
-    fail(`normalizeHubJob full row = ${JSON.stringify(full)}`);
-  }
-
-  // location assembled from locality/country, "Remote" appended when isRemote.
-  const assembled = normalizeHubJob({ title: 'R', absoluteJobUrl: 'https://thehub.io/jobs/r', location: { locality: 'Berlin', country: 'Germany' }, isRemote: true }, 'X');
-  if (assembled?.location === 'Berlin, Germany, Remote') pass('normalizeHubJob assembles locality/country and appends "Remote" when isRemote');
-  else fail(`normalizeHubJob assembled location = ${JSON.stringify(assembled?.location)}`);
-
-  // company fallbacks: entry name, then "The Hub".
-  const coEntry = normalizeHubJob({ title: 'T', absoluteJobUrl: 'https://thehub.io/jobs/c1', company: {} }, 'Entry Name');
-  const coDefault = normalizeHubJob({ title: 'T', absoluteJobUrl: 'https://thehub.io/jobs/c2' });
-  const coBlank = normalizeHubJob({ title: 'T', absoluteJobUrl: 'https://thehub.io/jobs/c3' }, '   '); // whitespace-only → "The Hub"
-  if (coEntry?.company === 'Entry Name' && coDefault?.company === 'The Hub' && coBlank?.company === 'The Hub') {
-    pass('normalizeHubJob falls back company → entry name → "The Hub" (whitespace-only entry name ignored)');
-  } else {
-    fail(`normalizeHubJob company fallbacks = ${JSON.stringify({ a: coEntry?.company, b: coDefault?.company, c: coBlank?.company })}`);
-  }
-
-  // postedAt falls back to createdAt; omitted when both absent.
-  const fromCreated = normalizeHubJob({ title: 'T', absoluteJobUrl: 'https://thehub.io/jobs/cd', createdAt: '2026-06-01T00:00:00.000Z' });
-  const noDate = normalizeHubJob({ title: 'T', absoluteJobUrl: 'https://thehub.io/jobs/nd' });
-  if (fromCreated?.postedAt === Date.parse('2026-06-01T00:00:00.000Z') && noDate && !('postedAt' in noDate)) {
-    pass('normalizeHubJob falls back to createdAt and omits postedAt when both dates are absent');
-  } else {
-    fail(`normalizeHubJob date handling = ${JSON.stringify({ created: fromCreated?.postedAt, none: noDate })}`);
-  }
-
-  // host-lock + drops: off-host url, non-https url, missing url, empty title, non-object.
-  const drops = [
-    normalizeHubJob({ title: 'Off host', absoluteJobUrl: 'https://evil.example/jobs/x' }),
-    normalizeHubJob({ title: 'Insecure', absoluteJobUrl: 'http://thehub.io/jobs/x' }),
-    normalizeHubJob({ title: 'No URL' }),
-    normalizeHubJob({ title: '', absoluteJobUrl: 'https://thehub.io/jobs/x' }),
-    normalizeHubJob(null),
-  ];
-  if (drops.every(r => r === null)) {
-    pass('normalizeHubJob host-locks absoluteJobUrl to thehub.io and drops off-host/non-https/no-url/empty-title/non-object');
-  } else {
-    fail(`normalizeHubJob drops = ${JSON.stringify(drops)}`);
-  }
-
-  // fetch(): pagination by ?page=N, stop on a short page.
-  const mk = (i) => ({ title: `Role ${i}`, absoluteJobUrl: `https://thehub.io/jobs/x${i}`, company: { name: `Co ${i}` }, location: { address: 'Copenhagen, Denmark' }, publishedAt: '2026-06-02T00:00:00.000Z' });
-  const hubPage1 = { docs: Array.from({ length: 15 }, (_, i) => mk(i)), page: 1, pages: 3, total: 33, limit: 15 };
-  const hubPage2 = { docs: [mk(15), mk(16), { title: '', absoluteJobUrl: 'https://thehub.io/jobs/bad' }], page: 2, pages: 3, limit: 15 }; // short → stop; 1 drop
-  const requested = [];
-  const pagedFetch = async (url, opts) => {
-    requested.push({ url, redirect: opts?.redirect });
-    const page = Number(new URL(url).searchParams.get('page'));
-    return page === 1 ? hubPage1 : hubPage2;
-  };
-  const paged = await thehub.fetch({ name: 'The Hub' }, { fetchJson: pagedFetch });
-
-  if (requested.length === 2
-      && requested[0].url === 'https://thehub.io/api/jobs?page=1'
-      && requested[1].url === 'https://thehub.io/api/jobs?page=2') {
-    pass('thehub.fetch() builds ?page=N URLs and stops after the short page');
-  } else {
-    fail(`thehub.fetch() requested = ${JSON.stringify(requested.map(r => r.url))}`);
-  }
-
-  if (requested.every(r => r.redirect === 'error')) pass('thehub.fetch() passes redirect:"error" on every page (SSRF guard)');
-  else fail(`thehub.fetch() redirect opts = ${JSON.stringify(requested.map(r => r.redirect))}`);
-
-  if (paged.length === 17) pass('thehub.fetch() aggregates valid jobs across pages (15 + 2, dropping the empty-title row)');
-  else fail(`thehub.fetch() returned ${paged.length} jobs (expected 17)`);
-
-  // Stops at the reported total pages even when every page is full.
-  const fullReq = [];
-  const fullPage = (page) => ({ docs: Array.from({ length: 15 }, (_, i) => mk(page * 100 + i)), page, pages: 2, limit: 15 });
-  await thehub.fetch(
-    { name: 'The Hub', max_pages: 10 },
-    { fetchJson: async (url) => { const p = Number(new URL(url).searchParams.get('page')); fullReq.push(p); return fullPage(p); } },
-  );
-  if (fullReq.length === 2 && fullReq[0] === 1 && fullReq[1] === 2) {
-    pass('thehub.fetch() stops at the reported total pages (pages:2) even with a higher max_pages');
-  } else {
-    fail(`thehub.fetch() pages-stop requested ${JSON.stringify(fullReq)}`);
-  }
-
-  // max_pages cap: only the first page is requested.
-  const capReq = [];
-  await thehub.fetch(
-    { name: 'The Hub', max_pages: 1 },
-    { fetchJson: async (url) => { capReq.push(url); return { docs: Array.from({ length: 15 }, (_, i) => mk(i)), page: 1, pages: 67, limit: 15 }; } },
-  );
-  if (capReq.length === 1 && capReq[0] === 'https://thehub.io/api/jobs?page=1') {
-    pass('thehub.fetch() honors max_pages (stops at the cap even on a full page)');
-  } else {
-    fail(`thehub.fetch() max_pages:1 requested ${JSON.stringify(capReq)}`);
-  }
-
-  // A full page with a large `pages` (so neither short-stop nor pages-stop fires)
-  // lets us assert the implicit default cap and the hard cap purely on page count.
-  const fullDeepPage = (page) => ({ docs: Array.from({ length: 15 }, (_, i) => mk(page * 100 + i)), page, pages: 999, limit: 15 });
-
-  // Default max_pages (no override) → exactly 3 pages.
-  const defReq = [];
-  await thehub.fetch(
-    { name: 'The Hub' },
-    { fetchJson: async (url) => { defReq.push(Number(new URL(url).searchParams.get('page'))); return fullDeepPage(defReq.length); } },
-  );
-  if (defReq.length === 3 && defReq[0] === 1 && defReq[2] === 3) {
-    pass('thehub.fetch() defaults to 3 pages when max_pages is not set');
-  } else {
-    fail(`thehub.fetch() default-cap requested ${JSON.stringify(defReq)} (expected pages 1..3)`);
-  }
-
-  // max_pages above the hard cap → clamped to 67 pages.
-  const cap67Req = [];
-  await thehub.fetch(
-    { name: 'The Hub', max_pages: 1000 },
-    { fetchJson: async (url) => { cap67Req.push(Number(new URL(url).searchParams.get('page'))); return fullDeepPage(cap67Req.length); } },
-  );
-  if (cap67Req.length === 67 && cap67Req[0] === 1 && cap67Req[66] === 67) {
-    pass('thehub.fetch() clamps max_pages to the hard cap of 67');
-  } else {
-    fail(`thehub.fetch() hard-cap requested ${cap67Req.length} pages (expected 67)`);
-  }
-
-  // unexpected API response → throws.
-  let badThrew = false;
-  try {
-    await thehub.fetch({ name: 'X' }, { fetchJson: async () => ({ wrong: true }) });
-  } catch (e) {
-    badThrew = /unexpected API response/.test(e.message);
-  }
-  if (badThrew) pass('thehub.fetch() throws on unexpected API response shape');
-  else fail('thehub.fetch() should throw when the docs array is absent');
-
-} catch (e) {
-  fail(`thehub provider tests crashed: ${e.message}`);
-}
-
-// ── 39. Provider — landingjobs ──────────────────────────────────
-console.log('\n39. Provider — landingjobs');
-
-try {
-  const ljModule = await import(pathToFileURL(join(ROOT, 'providers/landingjobs.mjs')).href);
-  const landingjobs = ljModule.default;
-  const { normalizeLandingJob, companyFromUrl } = ljModule;
-
-  if (landingjobs.id === 'landingjobs') pass('landingjobs.id is "landingjobs"');
-  else fail(`landingjobs.id is ${JSON.stringify(landingjobs.id)}`);
-
-  // companyFromUrl — humanizes the /at/<slug>/ segment; '' for other shapes.
-  if (companyFromUrl('https://landing.jobs/at/damia-group-portugal/some-job') === 'Damia Group Portugal'
-      && companyFromUrl('https://landing.jobs/at/inscale/x') === 'Inscale'
-      && companyFromUrl('https://landing.jobs/jobs/123') === ''
-      && companyFromUrl('not a url') === '') {
-    pass('companyFromUrl humanizes the /at/<slug>/ segment and returns "" for other shapes');
-  } else {
-    fail(`companyFromUrl = ${JSON.stringify([
-      companyFromUrl('https://landing.jobs/at/damia-group-portugal/some-job'),
-      companyFromUrl('https://landing.jobs/jobs/123'),
-    ])}`);
-  }
-
-  // normalizeLandingJob — full mapping.
-  const full = normalizeLandingJob(
-    { title: '  Senior Java Dev  ', url: 'https://landing.jobs/at/inscale/senior-java-dev', locations: [{ city: 'Lisbon', country_code: 'PT' }], remote: false, published_at: '2025-02-26T09:38:38.127Z' },
-    'Fallback',
-  );
-  if (full && full.title === 'Senior Java Dev' && full.url === 'https://landing.jobs/at/inscale/senior-java-dev'
-      && full.company === 'Inscale' && full.location === 'Lisbon, PT'
-      && full.postedAt === Date.parse('2025-02-26T09:38:38.127Z')) {
-    pass('normalizeLandingJob maps title/url, derives company from slug, builds location, parses published_at');
-  } else {
-    fail(`normalizeLandingJob full row = ${JSON.stringify(full)}`);
-  }
-
-  // remote:true appends "Remote".
-  const remoteJob = normalizeLandingJob({ title: 'R', url: 'https://landing.jobs/at/acme/r', locations: [{ city: 'Berlin', country_code: 'DE' }], remote: true });
-  if (remoteJob?.location === 'Berlin, DE, Remote') pass('normalizeLandingJob appends "Remote" when remote is true');
-  else fail(`normalizeLandingJob remote location = ${JSON.stringify(remoteJob?.location)}`);
-
-  // company fallback: url is landing.jobs but not /at/<slug>/ → entry name, then "Landing.jobs".
-  const coEntry = normalizeLandingJob({ title: 'T', url: 'https://landing.jobs/jobs/99' }, 'Entry Name');
-  const coDefault = normalizeLandingJob({ title: 'T', url: 'https://landing.jobs/jobs/100' });
-  const coBlank = normalizeLandingJob({ title: 'T', url: 'https://landing.jobs/jobs/101' }, '   '); // whitespace-only → "Landing.jobs"
-  if (coEntry?.company === 'Entry Name' && coDefault?.company === 'Landing.jobs' && coBlank?.company === 'Landing.jobs') {
-    pass('normalizeLandingJob falls back company → entry name → "Landing.jobs" (whitespace-only entry name ignored)');
-  } else {
-    fail(`normalizeLandingJob company fallbacks = ${JSON.stringify({ a: coEntry?.company, b: coDefault?.company, c: coBlank?.company })}`);
-  }
-
-  // postedAt falls back to created_at; omitted when both absent.
-  const fromCreated = normalizeLandingJob({ title: 'T', url: 'https://landing.jobs/at/acme/cd', created_at: '2025-02-26T08:53:02.254Z' });
-  const noDate = normalizeLandingJob({ title: 'T', url: 'https://landing.jobs/at/acme/nd' });
-  if (fromCreated?.postedAt === Date.parse('2025-02-26T08:53:02.254Z') && noDate && !('postedAt' in noDate)) {
-    pass('normalizeLandingJob falls back to created_at and omits postedAt when both dates are absent');
-  } else {
-    fail(`normalizeLandingJob date handling = ${JSON.stringify({ created: fromCreated?.postedAt, none: noDate })}`);
-  }
-
-  // host-lock + drops: off-host, non-https, missing url, empty title, non-object.
-  const drops = [
-    normalizeLandingJob({ title: 'Off host', url: 'https://evil.example/at/acme/x' }),
-    normalizeLandingJob({ title: 'Insecure', url: 'http://landing.jobs/at/acme/x' }),
-    normalizeLandingJob({ title: 'No URL' }),
-    normalizeLandingJob({ title: '', url: 'https://landing.jobs/at/acme/x' }),
-    normalizeLandingJob(null),
-  ];
-  if (drops.every(r => r === null)) {
-    pass('normalizeLandingJob host-locks url to landing.jobs and drops off-host/non-https/no-url/empty-title/non-object');
-  } else {
-    fail(`normalizeLandingJob drops = ${JSON.stringify(drops)}`);
-  }
-
-  // fetch(): single call to the feed URL, normalized, with the SSRF guard.
-  const sample = [
-    { title: 'Role A', url: 'https://landing.jobs/at/acme/role-a', locations: [{ city: 'Porto', country_code: 'PT' }], remote: false, published_at: '2025-02-26T09:38:38.127Z' },
-    { title: '', url: 'https://landing.jobs/at/acme/bad' }, // dropped: empty title
-  ];
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await landingjobs.fetch(
-    { name: 'Landing.jobs' },
-    { fetchJson: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://landing.jobs/api/v1/jobs') pass('landingjobs.fetch() requests the v1 feed URL');
-  else fail(`landingjobs.fetch() requested ${JSON.stringify(capturedUrl)}`);
-
-  if (capturedOpts && capturedOpts.redirect === 'error') pass('landingjobs.fetch() passes redirect:"error" (SSRF guard)');
-  else fail(`landingjobs.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-
-  if (fetched.length === 1 && fetched[0]?.company === 'Acme') {
-    pass('landingjobs.fetch() returns normalized jobs (drops the empty-title row, derives company)');
-  } else {
-    fail(`landingjobs.fetch() returned ${fetched.length} jobs, row 0 = ${JSON.stringify(fetched[0])}`);
-  }
-
-  // unexpected (non-array) response → throws.
-  let badThrew = false;
-  try {
-    await landingjobs.fetch({ name: 'X' }, { fetchJson: async () => ({ jobs: [] }) });
-  } catch (e) {
-    badThrew = /unexpected API response/.test(e.message);
-  }
-  if (badThrew) pass('landingjobs.fetch() throws on a non-array API response');
-  else fail('landingjobs.fetch() should throw when the response is not an array');
-
-} catch (e) {
-  fail(`landingjobs provider tests crashed: ${e.message}`);
-}
-  
-
-// -- 40. Provider - hackernews ---------------------------------------
-console.log('\n40. Provider - hackernews');
-
-try {
-  const hn = (await import(pathToFileURL(join(ROOT, 'providers/hackernews.mjs')).href)).default;
-  const { parseHnComment, resolveLatestThreadId } =
-    await import(pathToFileURL(join(ROOT, 'providers/hackernews.mjs')).href);
-
-  // resolveLatestThreadId ─ happy path
-  const fakeSearch = {
-    hits: [
-      { objectID: '99999999', title: 'Ask HN: Who is hiring? (June 2025)' },
-      { objectID: '88888888', title: 'Ask HN: Who wants to be hired? (June 2025)' },
-    ],
-  };
-  if (resolveLatestThreadId(fakeSearch) === '99999999') {
-    pass('resolveLatestThreadId picks the first matching "Who is hiring?" hit');
-  } else {
-    fail(`resolveLatestThreadId returned ${resolveLatestThreadId(fakeSearch)}`);
-  }
-
-  // resolveLatestThreadId ─ no matching hit
-  const noMatch = { hits: [{ objectID: '11111', title: 'Ask HN: Who wants to be hired?' }] };
-  if (resolveLatestThreadId(noMatch) === null) {
-    pass('resolveLatestThreadId returns null when no thread matches');
-  } else {
-    fail('resolveLatestThreadId should return null for non-hiring threads');
-  }
-
-  // resolveLatestThreadId ─ bad input
-  if (resolveLatestThreadId(null) === null && resolveLatestThreadId({}) === null) {
-    pass('resolveLatestThreadId handles null / empty input gracefully');
-  } else {
-    fail('resolveLatestThreadId should return null for bad input');
-  }
-
-  // parseHnComment ─ full pipe-delimited format
-  const fullFmt = 'Acme Corp | Senior Engineer | Remote | https://acme.com/jobs/123\nWe are hiring…';
-  const parsed = parseHnComment(fullFmt, 'https://news.ycombinator.com/item?id=1');
-  if (parsed && parsed.company === 'Acme Corp' && parsed.location === 'Remote') {
-    pass('parseHnComment extracts company and location from pipe-delimited header');
-  } else {
-    fail(`parseHnComment pipe format: ${JSON.stringify(parsed)}`);
-  }
-  if (parsed && parsed.url === 'https://acme.com/jobs/123') {
-    pass('parseHnComment extracts URL from comment text');
-  } else {
-    fail(`parseHnComment url: ${JSON.stringify(parsed?.url)}`);
-  }
-
-  // parseHnComment ─ URL in first line stripped from title
-  if (parsed && !parsed.title.includes('https://')) {
-    pass('parseHnComment strips URLs from the title field');
-  } else {
-    fail(`parseHnComment title still contains URL: ${JSON.stringify(parsed?.title)}`);
-  }
-
-  // parseHnComment ─ free-form (no pipes, no URL in text)
-  const freeFmt = 'Looking for a Rails dev, anywhere, part-time.';
-  const parsedFree = parseHnComment(freeFmt, 'https://news.ycombinator.com/item?id=1');
-  if (parsedFree && parsedFree.title === freeFmt && parsedFree.url === 'https://news.ycombinator.com/item?id=1') {
-    pass('parseHnComment handles free-form comment, falls back to thread URL');
-  } else {
-    fail(`parseHnComment free-form: ${JSON.stringify(parsedFree)}`);
-  }
-  if (parsedFree && parsedFree.company === '' && parsedFree.location === '') {
-    pass('parseHnComment leaves company/location empty for free-form comments');
-  } else {
-    fail(`parseHnComment free-form company/location: ${JSON.stringify(parsedFree)}`);
-  }
-
-  // parseHnComment ─ HTML entities and tags
-  const htmlFmt = '<p>Beta &amp; Co | Staff SWE | New York, NY | <a href="https://beta.io/jobs">https://beta.io/jobs</a></p>';
-  const parsedHtml = parseHnComment(htmlFmt, '');
-  if (parsedHtml && parsedHtml.company === 'Beta & Co') {
-    pass('parseHnComment decodes HTML entities (company name)');
-  } else {
-    fail(`parseHnComment HTML entity decode: ${JSON.stringify(parsedHtml?.company)}`);
-  }
-  if (parsedHtml && parsedHtml.url === 'https://beta.io/jobs') {
-    pass('parseHnComment extracts href URL from anchor tags');
-  } else {
-    fail(`parseHnComment anchor URL: ${JSON.stringify(parsedHtml?.url)}`);
-  }
-
-  // parseHnComment ─ <p> paragraph body must not bleed into location field
-  // Real HN posts often have "Company | Role | Location<p>Body text..." where the
-  // second <p> paragraph (job description) runs on without a 4th pipe segment.
-  // The parser must convert block tags to newlines so parts[2] stays clean.
-  const bleedFmt = '<p>Linear | Product Engineer | Remote<p>We are hiring at https://linear.app/careers/pe-2025';
-  const parsedBleed = parseHnComment(bleedFmt, 'https://news.ycombinator.com/item?id=1');
-  if (parsedBleed && parsedBleed.location === 'Remote') {
-    pass('parseHnComment location does not bleed body paragraph text (block tag newline fix)');
-  } else {
-    fail(`parseHnComment location bleed: ${JSON.stringify(parsedBleed?.location)}`);
-  }
-  if (parsedBleed && parsedBleed.url === 'https://linear.app/careers/pe-2025') {
-    pass('parseHnComment finds URL in body paragraph when absent from header line');
-  } else {
-    fail(`parseHnComment body-paragraph URL: ${JSON.stringify(parsedBleed?.url)}`);
-  }
-
-  // parseHnComment ─ deleted / empty comments return null
-  if (parseHnComment('', '') === null && parseHnComment(null, '') === null) {
-    pass('parseHnComment returns null for empty / null input');
-  } else {
-    fail('parseHnComment should return null for empty/null input');
-  }
-
-  // provider.fetch() — integration with mock ctx
-  const FAKE_THREAD_ID = '42424242';
-  const FAKE_THREAD_URL = `https://news.ycombinator.com/item?id=${FAKE_THREAD_ID}`;
-
-  const fakeSearchResp = {
-    hits: [{ objectID: FAKE_THREAD_ID, title: 'Ask HN: Who is hiring? (June 2025)' }],
-  };
-  const fakeItemResp = {
-    id: FAKE_THREAD_ID,
-    children: [
-      {
-        objectID: 'c1',
-        text: 'Startup XYZ | Backend Engineer | San Francisco | https://xyz.io/careers',
-        created_at: '2025-06-01T10:00:00Z',
-      },
-      { objectID: 'c2', deleted: true, text: '' },      // deleted — should be skipped
-      { objectID: 'c3', text: '' },                      // empty — should be skipped
-      {
-        objectID: 'c4',
-        text: 'Freelance gig, no URL here, DM me.',      // no URL — falls back to thread URL
-        created_at: '2025-06-01T11:00:00Z',
-      },
-    ],
-  };
-
-  let searchFetched = false;
-  let itemFetched = false;
-  const mockCtx = {
-    async fetchJson(url, _opts) {
-      if (url.includes('search_by_date')) { searchFetched = true; return fakeSearchResp; }
-      if (url.includes(`/items/${FAKE_THREAD_ID}`)) { itemFetched = true; return fakeItemResp; }
-      throw new Error(`hackernews mock: unexpected fetch ${url}`);
-    },
-  };
-
-  const jobs = await hn.fetch({ name: 'HN Hiring', provider: 'hackernews' }, mockCtx);
-
-  if (searchFetched && itemFetched) {
-    pass('hackernews.fetch() calls search API then items API');
-  } else {
-    fail(`hackernews.fetch() API calls: search=${searchFetched} item=${itemFetched}`);
-  }
-
-  if (jobs.length === 2) {
-    pass('hackernews.fetch() returns 2 jobs (skips deleted + empty comments)');
-  } else {
-    fail(`hackernews.fetch() returned ${jobs.length} jobs (expected 2)`);
-  }
-
-  if (jobs[0]?.company === 'Startup XYZ' && jobs[0]?.location === 'San Francisco') {
-    pass('hackernews.fetch() maps pipe-delimited company and location');
-  } else {
-    fail(`hackernews.fetch() row 0: ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[0]?.url === 'https://xyz.io/careers') {
-    pass('hackernews.fetch() maps job URL from comment text');
-  } else {
-    fail(`hackernews.fetch() row 0 url: ${JSON.stringify(jobs[0]?.url)}`);
-  }
-
-  if (jobs[0]?.postedAt === Date.parse('2025-06-01T10:00:00Z')) {
-    pass('hackernews.fetch() parses created_at to postedAt epoch ms');
-  } else {
-    fail(`hackernews.fetch() postedAt: ${JSON.stringify(jobs[0]?.postedAt)}`);
-  }
-
-  if (jobs[1]?.url === FAKE_THREAD_URL) {
-    pass('hackernews.fetch() falls back to thread URL for comments with no link');
-  } else {
-    fail(`hackernews.fetch() fallback url: ${JSON.stringify(jobs[1]?.url)}`);
-  }
-
-  // provider.fetch() — throws when no thread found
-  let threw = false;
-  try {
-    await hn.fetch({ name: 'HN' }, {
-      async fetchJson() { return { hits: [] }; },
-    });
-  } catch (e) {
-    threw = true;
-  }
-  if (threw) {
-    pass('hackernews.fetch() throws when "Who is hiring?" thread not found');
-  } else {
-    fail('hackernews.fetch() should throw when no matching thread found');
-  }
-} catch (e) {
-  fail(`hackernews provider tests crashed: ${e.message}`);
-}
-
-// -- 40. Provider - himalayas --------------------------------------------
-console.log('\n40. Provider - himalayas');
-
-try {
-  const himalayasModule = await import(pathToFileURL(join(ROOT, 'providers/himalayas.mjs')).href);
-  const himalayas = himalayasModule.default;
-  const { parseHimalayasResponse } = himalayasModule;
-
-  if (himalayas.id === 'himalayas') pass('himalayas.id is "himalayas"');
-  else fail(`himalayas.id is ${JSON.stringify(himalayas.id)}`);
-
-  const hit = himalayas.detect({ name: 'Himalayas', provider: 'himalayas' });
-  if (hit && hit.url === 'https://himalayas.app/jobs/api?limit=50') {
-    pass('himalayas.detect() claims explicit provider config');
-  } else {
-    fail(`himalayas.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  if (himalayas.detect({ name: 'Remote Board', provider: 'remotive' }) === null) {
-    pass('himalayas.detect() ignores other provider ids');
-  } else {
-    fail('himalayas.detect() should only claim provider: himalayas');
-  }
-
-  const sample = {
-    jobs: [
-      {
-        title: '  Staff AI Engineer  ',
-        companyName: ' Acme Labs ',
-        companySlug: 'acme-labs',
-        locationRestrictions: ['Worldwide', 'Europe'],
-        pubDate: 1782538666,
-        applicationLink: 'https://himalayas.app/companies/acme-labs/jobs/staff-ai-engineer',
-        guid: 'https://himalayas.app/companies/acme-labs/jobs/staff-ai-engineer-guid',
-      },
-      {
-        title: 'Product Manager',
-        companyName: 'Fallback Co',
-        companySlug: 'fallback-co',
-        locationRestrictions: [],
-        pubDate: '2026-01-02T09:00:00Z',
-        applicationLink: '',
-        guid: 'https://himalayas.app/companies/fallback-co/jobs/product-manager',
-      },
-      {
-        title: 'Missing Link Role',
-        companyName: 'Dropped Co',
-        locationRestrictions: ['United States'],
-      },
-      {
-        title: 'Off Host Role',
-        companyName: 'Bad Co',
-        locationRestrictions: ['Remote'],
-        applicationLink: 'https://example.com/companies/bad/jobs/off-host',
-      },
-      {
-        title: 'HTTP Role',
-        companyName: 'Bad Scheme Co',
-        locationRestrictions: ['Remote'],
-        applicationLink: 'http://himalayas.app/companies/bad/jobs/http-role',
-      },
-      {
-        title: '   ',
-        companyName: 'Blank Title Co',
-        locationRestrictions: ['Remote'],
-        applicationLink: 'https://himalayas.app/companies/blank/jobs/blank-title',
-      },
-    ],
-  };
-  const jobs = parseHimalayasResponse(sample);
-
-  if (jobs.length === 2) pass('parseHimalayasResponse keeps 2 jobs (drops missing/off-host/http/blank-title rows)');
-  else fail(`parseHimalayasResponse returned ${jobs.length} jobs (expected 2)`);
-
-  if (jobs[0]?.title === 'Staff AI Engineer' && jobs[0]?.company === 'Acme Labs') {
-    pass('parseHimalayasResponse trims title and companyName');
-  } else {
-    fail(`row 0 title/company = ${JSON.stringify({ title: jobs[0]?.title, company: jobs[0]?.company })}`);
-  }
-
-  if (jobs[0]?.location === 'Worldwide, Europe') {
-    pass('parseHimalayasResponse joins locationRestrictions');
-  } else {
-    fail(`row 0 location = ${JSON.stringify(jobs[0]?.location)}`);
-  }
-
-  if (jobs[0]?.url === 'https://himalayas.app/companies/acme-labs/jobs/staff-ai-engineer') {
-    pass('parseHimalayasResponse maps applicationLink to url');
-  } else {
-    fail(`row 0 url = ${JSON.stringify(jobs[0]?.url)}`);
-  }
-
-  if (jobs[0]?.postedAt === 1782538666 * 1000) {
-    pass('parseHimalayasResponse converts epoch seconds pubDate -> postedAt ms');
-  } else {
-    fail(`row 0 postedAt = ${JSON.stringify(jobs[0]?.postedAt)}`);
-  }
-
-  if (jobs[1]?.url === 'https://himalayas.app/companies/fallback-co/jobs/product-manager') {
-    pass('parseHimalayasResponse falls back to guid when applicationLink is missing');
-  } else {
-    fail(`row 1 url = ${JSON.stringify(jobs[1]?.url)}`);
-  }
-
-  if (jobs[1]?.postedAt === Date.parse('2026-01-02T09:00:00Z')) {
-    pass('parseHimalayasResponse parses string pubDate -> postedAt');
-  } else {
-    fail(`row 1 postedAt = ${JSON.stringify(jobs[1]?.postedAt)}`);
-  }
-
-  if (parseHimalayasResponse({}).length === 0 && parseHimalayasResponse(null).length === 0) {
-    pass('parseHimalayasResponse empty / non-object payload -> empty result (no crash)');
-  } else {
-    fail('parseHimalayasResponse invalid payload should yield empty result');
-  }
-
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await himalayas.fetch(
-    { name: 'Himalayas', provider: 'himalayas' },
-    { fetchJson: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://himalayas.app/jobs/api?limit=50') {
-    pass('himalayas.fetch() requests the pinned API URL');
-  } else {
-    fail(`himalayas.fetch() requested ${JSON.stringify(capturedUrl)}`);
-  }
-
-  if (capturedOpts && capturedOpts.redirect === 'error') {
-    pass('himalayas.fetch() passes redirect:"error" to fetchJson');
-  } else {
-    fail(`himalayas.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-  }
-
-  if (fetched[0]?.company === 'Acme Labs' && fetched[0]?.title === 'Staff AI Engineer') {
-    pass('provider: himalayas config returns normalized jobs');
-  } else {
-    fail(`himalayas.fetch() normalized row = ${JSON.stringify(fetched[0])}`);
-  }
-} catch (e) {
-  fail(`himalayas provider tests crashed: ${e.message}`);
-}
-
-// ── 41. Provider — jobicy ───────────────────────────────────────
-console.log('\n41. Provider — jobicy');
-
-try {
-  const jobicy = (await import(pathToFileURL(join(ROOT, 'providers/jobicy.mjs')).href)).default;
-  const { parseJobicyResponse } = await import(pathToFileURL(join(ROOT, 'providers/jobicy.mjs')).href);
-
-  if (jobicy.id === 'jobicy') pass('jobicy.id is "jobicy"');
-  else fail(`jobicy.id is ${JSON.stringify(jobicy.id)}`);
-
-  const hit = jobicy.detect({ name: 'Jobicy Board', provider: 'jobicy' });
-  if (hit && hit.url === 'https://jobicy.com/api/v2/remote-jobs?count=50') {
-    pass('jobicy.detect() claims explicit provider config');
-  } else {
-    fail(`jobicy.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  if (jobicy.detect({ name: 'Remote Board', provider: 'remoteok' }) === null) {
-    pass('jobicy.detect() ignores other provider ids');
-  } else {
-    fail('jobicy.detect() should only claim provider: jobicy');
-  }
-
-  const sample = {
-    jobs: [
-      {
-        jobTitle: 'Senior AI Engineer',
-        companyName: 'Acme Corp',
-        jobGeo: 'Worldwide',
-        url: 'https://jobicy.com/jobs/senior-ai-engineer',
-        pubDate: '2026-06-27T10:00:00',
-      },
-      {
-        jobTitle: 'Staff Backend Developer',
-        companyName: 'Globex',
-        jobGeo: 'Europe',
-        url: 'https://jobicy.com/jobs/staff-backend-developer',
-        pubDate: '2026-06-25T12:00:00Z',
-      },
-      {
-        jobTitle: 'Role With Missing URL',
-        companyName: 'Incomplete',
-        jobGeo: 'USA',
-        pubDate: '2026-06-24T08:00:00Z',
-      },
-      {
-        jobTitle: 'Role With Invalid URL',
-        companyName: 'Invalid',
-        url: 'not-a-valid-url',
-        jobGeo: 'USA',
-      },
-      {
-        jobTitle: '',
-        companyName: 'Empty Title',
-        url: 'https://jobicy.com/jobs/empty-title',
-        jobGeo: 'USA',
-      }
-    ]
-  };
-
-  const jobs = parseJobicyResponse(sample, 'Jobicy Board');
-
-  if (jobs.length === 2) pass('parseJobicyResponse keeps 2 jobs (drops missing/invalid url and empty title)');
-  else fail(`parseJobicyResponse returned ${jobs.length} jobs (expected 2)`);
-
-  if (jobs[0]?.company === 'Acme Corp' && jobs[0]?.title === 'Senior AI Engineer') {
-    pass('parseJobicyResponse maps jobTitle -> title, companyName -> company');
-  } else {
-    fail(`row 0 title/company = ${JSON.stringify({ title: jobs[0]?.title, company: jobs[0]?.company })}`);
-  }
-
-  if (jobs[0]?.url === 'https://jobicy.com/jobs/senior-ai-engineer') {
-    pass('parseJobicyResponse maps url to url');
-  } else {
-    fail(`row 0 url = ${JSON.stringify(jobs[0]?.url)}`);
-  }
-
-  if (jobs[0]?.location === 'Worldwide') {
-    pass('parseJobicyResponse maps jobGeo to location');
-  } else {
-    fail(`row 0 location = ${JSON.stringify(jobs[0]?.location)}`);
-  }
-
-  if (jobs[0]?.postedAt === Date.parse('2026-06-27T10:00:00')) {
-    pass('parseJobicyResponse parses pubDate -> postedAt');
-  } else {
-    fail(`row 0 postedAt = ${JSON.stringify(jobs[0]?.postedAt)}`);
-  }
-
-  if (jobs[1]?.company === 'Globex' && jobs[1]?.title === 'Staff Backend Developer') {
-    pass('parseJobicyResponse parses second job correctly');
-  } else {
-    fail(`row 1 = ${JSON.stringify(jobs[1])}`);
-  }
-
-  if (parseJobicyResponse('', 'X').length === 0 && parseJobicyResponse(null, 'X').length === 0) {
-    pass('parseJobicyResponse empty / non-object payload -> empty result (no crash)');
-  } else {
-    fail('parseJobicyResponse empty / non-object payload should yield empty result');
-  }
-
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await jobicy.fetch(
-    { name: 'Jobicy Board', provider: 'jobicy' },
-    { fetchJson: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://jobicy.com/api/v2/remote-jobs?count=50') {
-    pass('jobicy.fetch() requests the pinned JSON feed URL');
-  } else {
-    fail(`jobicy.fetch() requested ${JSON.stringify(capturedUrl)}`);
-  }
-
-  if (capturedOpts && capturedOpts.redirect === 'error') {
-    pass('jobicy.fetch() passes redirect:"error" to fetchJson');
-  } else {
-    fail(`jobicy.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-  }
-
-  if (fetched[0]?.company === 'Acme Corp' && fetched[0]?.title === 'Senior AI Engineer') {
-    pass('provider: jobicy config returns normalized jobs');
-  } else {
-    fail(`jobicy.fetch() normalized row = ${JSON.stringify(fetched[0])}`);
-  }
-
-} catch (e) {
-  fail(`jobicy provider tests crashed: ${e.message}`);
-}
-
-// ── 42. PROVIDERS — JustJoin.it ───────────────────────────────────────────
-
-console.log('\n42. Provider — justjoin');
-
-try {
-  const jj = (await import(pathToFileURL(join(ROOT, 'providers/justjoin.mjs')).href)).default;
-  const { parseJustJoinResponse } = await import(pathToFileURL(join(ROOT, 'providers/justjoin.mjs')).href);
-
-  if (jj.id === 'justjoin') pass('justjoin.id is "justjoin"');
-  else fail(`justjoin.id is ${JSON.stringify(jj.id)}`);
-
-  if (jj.detect({ name: 'JustJoin', careers_url: 'https://justjoin.it/job-offers/all-locations' })?.url) {
-    pass('justjoin.detect() matches job-offers URL');
-  } else {
-    fail('justjoin.detect() should match justjoin.it job-offers URL');
-  }
-
-  if (jj.detect({ name: 'X', careers_url: 'https://evil.example/justjoin.it/job-offers' }) === null) {
-    pass('justjoin.detect() rejects path-spoofed URLs');
-  } else {
-    fail('justjoin.detect() must reject path-spoofed URLs');
-  }
-
-  if (jj.detect({ name: 'X', api: 'https://justjoin.it/api/candidate-api/offers/count' }) === null) {
-    pass('justjoin.detect() rejects non-offers API paths');
-  } else {
-    fail('justjoin.detect() must reject non-offers API paths');
-  }
-
-  const fakeResponse = {
-    data: [
-      {
-        slug: 'acme-senior-dev-warsaw-javascript',
-        title: 'Senior Developer',
-        companyName: 'Acme',
-        workplaceType: 'remote',
-        locations: [{ city: 'Warszawa' }],
-        publishedAt: '2026-06-12T10:00:00.000Z',
-      },
-      { slug: '', title: 'Missing Slug', companyName: 'Broken' },
-      null,
-    ],
-    meta: { next: { cursor: null } },
-  };
-
-  const parsed = parseJustJoinResponse(fakeResponse);
-  if (parsed.length === 1) pass('justjoin parser filters malformed rows');
-  else fail(`justjoin parser returned ${parsed.length} rows, expected 1`);
-
-  if (parsed[0].title === 'Senior Developer' && parsed[0].url === 'https://justjoin.it/job-offer/acme-senior-dev-warsaw-javascript') {
-    pass('justjoin parser maps title and URL');
-  } else {
-    fail(`justjoin parser mapped title/url incorrectly: ${JSON.stringify(parsed[0])}`);
-  }
-
-  if (parsed[0].company === 'Acme' && parsed[0].location === 'remote, Warszawa' && typeof parsed[0].postedAt === 'number') {
-    pass('justjoin parser maps company, location, and postedAt');
-  } else {
-    fail(`justjoin parser mapped fields incorrectly: ${JSON.stringify(parsed[0])}`);
-  }
-
-  let capturedUrl = '';
-  let capturedOpts = null;
-  const fetched = await jj.fetch(
-    { name: 'JustJoin', careers_url: 'https://justjoin.it/job-offers/all-locations', max_pages: 1 },
-    {
-      transport: 'http',
-      fetchJson: async (url, opts) => {
-        capturedUrl = url;
-        capturedOpts = opts;
-        return fakeResponse;
-      },
-      fetchText: async () => '',
-    },
-  );
-  if (fetched.length === 1 && capturedUrl.startsWith('https://justjoin.it/api/candidate-api/offers?')) {
-    pass('justjoin.fetch() uses candidate API endpoint');
-  } else {
-    fail(`justjoin.fetch() endpoint/result wrong: ${capturedUrl} ${JSON.stringify(fetched)}`);
-  }
-
-  if (capturedOpts && capturedOpts.redirect === 'error') pass('justjoin.fetch() passes redirect:"error"');
-  else fail(`justjoin.fetch() should pass redirect:"error", got ${JSON.stringify(capturedOpts)}`);
-
-  let ssrfRejected = false;
-  try {
-    await jj.fetch(
-      { name: 'Evil', careers_url: 'https://evil.example/job-offers/all-locations' },
-      {
-        transport: 'http',
-        fetchJson: async () => { throw new Error('SSRF! should not reach here'); },
-        fetchText: async () => '',
-      },
-    );
-  } catch (e) {
-    if (e.message.includes('trusted justjoin.it')) ssrfRejected = true;
-  }
-  if (ssrfRejected) pass('justjoin.fetch() rejects untrusted host');
-  else fail('justjoin.fetch() should reject untrusted host');
-
-  let badShape = false;
-  try {
-    parseJustJoinResponse({ jobs: [] });
-  } catch (e) {
-    if (e.message.includes('unexpected API response')) badShape = true;
-  }
-  if (badShape) pass('justjoin parser throws on bad response shape');
-  else fail('justjoin parser should throw on bad response shape');
-} catch (e) {
-  fail(`justjoin provider tests crashed: ${e.message}`);
-}
-
 // ── 43. PROVIDERS — NoFluffJobs ───────────────────────────────────────────
 
 console.log('\n43. Provider — nofluffjobs');
